+++ src/locales/cn/ui.json
@@ -402,7 +402,6 @@
     "Select card to add 4 microbes": "选择需要添加 4 个微生物资源的卡",
     "Remove 1 floater to increase M€ production 1 step": "移除 1 个云资源提升 M€ 产能 1 级",
     "Draw 3 Prelude cards, and play one of them": "抽 3 张前序牌,打出其中1张",
-    "Choose prelude card to play": "选择要打出的前序牌",
     "Remove 1 floater on this card to trade for free": "移除 1 个云资源免费贸易 1 次",
     "Add 1 floater to a Jovian card": "添加 1 个云资源到任意木星卡",
     "Pay 1 floater (use ${0} action)": "支付 1 个云资源(使用 泰坦漂浮物起飞平台 行动)",
