+++ src/locales/de/ui.json
@@ -640,7 +640,6 @@
     "Draw 2 cards with a microbe tag": "Ziehe 2 Karten mit Mikroben-Symbol",
     "Spend 3 M€ to draw a blue card": "Bezahle 3 M€ um 1 blaue Karte zu ziehen",
     "Flip and discard a played blue card": "Dreh eine gespielte blaue Karte um und verwerfe sie",
-    "Choose prelude card to play": "Wähle eine Präludium-Karte um sie zu spielen",
     "Choose CEO card to play": "Wähle eine CEO-Karte um sie zu spielen",
     "Remove 1 floater on this card to trade for free": "Entferne 1 Schweber von dieser Karte, um kostenlos zu handeln",
     "Add 1 floater to a Jovian card": "Füge einer JUPITER-KARTE 1 Schweber hinzu",
