+++ src/locales/fr/prelude.json
@@ -1,5 +1,4 @@
 {
     "Increase your energy production 3 steps.": "Augmentez votre production d'énergie de 3.",
-    "Draw 3 Prelude cards, and play one of them": "Piocher 3 cartes prélude et jouer une d'entre elles",
-    "Choose prelude card to play": "Choisir la carte prélude à jouer"
+    "Draw 3 Prelude cards, and play one of them": "Piocher 3 cartes prélude et jouer une d'entre elles"
 }
