+++ src/locales/ko/UI.json
@@ -401,7 +401,6 @@
     "Select colony for Research Colony": "‘연구형 개척기지’ 카드로 건설할 개척기지의 위성 선택",
     "Remove 1 asteroid to raise Venus 1 step": "소행성 자원 1개를 소비하여 금성개척률 1단계 상승",
     "Pay 6 M€ to add 1 asteroid to this card": "6 M€을 지불하여 이 카드에 소행성 자원 1개 추가",
-    "Pay 6 M€ to add 1 asteroid to this card": "6 M€을 지불하여 이 카드에 소행성 자원 1개 추가",
     "Remove asteroid": "소행성 자원 소비",
     "Add a colony tile": "위성 타일 하나 추가",
     "ARIDOR": "아리도르",
@@ -412,7 +411,6 @@
     "Select card to add 4 microbes": "미생물 자원 4개를 추가할 카드 선택",
     "Remove 1 floater to increase M€ production 1 step": "부양체 자원 1개를 소비하여 메가크레딧 생산력 1 증가",
     "Draw 3 Prelude cards, and play one of them": "서곡 카드 3장을 뽑고 그 중 하나를 추진",
-    "Choose prelude card to play": "추진할 서곡 카드 선택",
     "Remove 1 floater on this card to trade for free": "이 카드의 부양체 자원 1개를 소비하여 무료로 무역사업 수행",
     "Add 1 floater to a Jovian card": "목성 카드에 부양체 자원 1개 추가",
     "Pay 1 floater (use ${0} action)": "부양체 자원 1개 지불 (‘타이탄 공중 발사대’ 행동)",
