+++ src/locales/nl/ui.json
@@ -549,7 +549,6 @@
     "Draw 2 cards with a microbe tag": "Trek 2 kaarten met een microbe-symbool",
     "Spend 3 M€ to draw a blue card": "Betaal 3 M€ om 1 blauwe kaart te trekken",
     "Flip and discard a played blue card": "Draai om en leg een gespeelde blauwe kaart af",
-    "Choose prelude card to play": "Kies een Prelude-kaart om te spelen",
     "Remove 1 floater on this card to trade for free": "Verwijder 1 floater van deze kaart om gratis te ruilen",
     "Add 1 floater to a Jovian card": "Voeg 1 floater toe aan een Jupiter-kaart",
     "Pay 1 floater (use ${0} action)": "Betaal 1 floater (gebruik de Titan drijvend lanceerplatform actie)",
