+++ tests/cards/prelude/ValleyTrust.spec.ts
@@ -1,23 +1,21 @@
 import {SelectCard} from '../../../src/server/inputs/SelectCard';
 import {expect} from 'chai';
-import {cast, runAllActions} from '../../TestingUtils';
+import {cast} from '../../TestingUtils';
 import {Ants} from '../../../src/server/cards/base/Ants';
 import {MedicalLab} from '../../../src/server/cards/base/MedicalLab';
 import {Research} from '../../../src/server/cards/base/Research';
 import {ValleyTrust} from '../../../src/server/cards/prelude/ValleyTrust';
 import {TestPlayer} from '../../TestPlayer';
 import {testGame} from '../../TestGame';
-import {IGame} from '../../../src/server/IGame';
 import {IPreludeCard, isPreludeCard} from '../../../src/server/cards/prelude/IPreludeCard';
 
 describe('ValleyTrust', function() {
-  let game: IGame;
   let card: ValleyTrust;
   let player: TestPlayer;
 
   beforeEach(function() {
     card = new ValleyTrust();
-    [game, player] = testGame(1, {preludeExtension: true});
+    [, player] = testGame(1, {preludeExtension: true});
   });
 
   it('Does not get card discount for other tags', function() {
@@ -34,19 +32,16 @@ describe('ValleyTrust', function() {
   });
 
   it('initial action', () => {
-    cast(card.initialAction(player), undefined);
-    runAllActions(game);
-    const selectCard = cast(player.popWaitingFor(), SelectCard<IPreludeCard>);
+    const selectCard = cast(card.initialAction(player), SelectCard<IPreludeCard>);
 
     expect(selectCard.cards).has.length(3);
     expect(selectCard.cards.every((c) => isPreludeCard(c))).is.true;
   });
 
   it('Card works even without prelude expansion enabled', () => {
-    [game, player] = testGame(1, {preludeExtension: false});
-    cast(card.initialAction(player), undefined);
-    runAllActions(game);
-    const selectCard = cast(player.popWaitingFor(), SelectCard<IPreludeCard>);
+    [, player] = testGame(1, {preludeExtension: false});
+    const selectCard = cast(card.initialAction(player), SelectCard<IPreludeCard>);
+
     expect(selectCard.cards).has.length(3);
     expect(selectCard.cards.every((c) => isPreludeCard(c))).is.true;
   });
